Safari-609.1.11.1/LayoutTests/editing/selection/select-out-of-floated-non-editable-08.html

<!DOCTYPE html>
<html>
<body>
<div id="container"><div>block-level text before floats</div><div id="target" style="float: left; margin-right: 20px">first float*</div><div style="float: left; margin-right: 20px">second float</div></div>
<p style="padding-top: 1em;">
To test manually, drag from the middle of the 'first-float' div to the
right, into the 'second float' element. The selection should end
inside the 'second float' element and not jump to the beginning.
</p>
<pre id="console"></pre>
<script src="../../resources/dump-as-markup.js"></script>
<script src="resources/select-out-of-floated-non-editable.js"></script>
<script>
runTest("float*s", 25);
</script>
</body>
</html>
