Safari-611.1.10/LayoutTests/editing/selection/select-out-of-floated-non-editable-08-expected.txt

block-level text before floats
first float*second float
To test manually, drag from the middle of the 'first-float' div to the right, into the 'second float' element. The selection should end inside the 'second float' element and not jump to the beginning.

SUCCESS: Selected test is 'float*s' as expected
| <div>
|   "block-level text before floats"
| <div>
|   id="target"
|   style="float: left; margin-right: 20px"
|   "first <#selection-anchor>float*"
| <div>
|   style="float: left; margin-right: 20px"
|   "s<#selection-focus>econd float"
